requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
Pillow==9.0.1
fcm-django==1.0.10
selenium==4.1.3
beautifulsoup4==4.10.0
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
whitenoise==6.0.0
pytest==7.1.1
chromedriver-autoinstaller==0.3.1
webdriver-manager==3.5.4
shot-scraper>=0.7